# docs/api/openapi.yml
openapi: 3.1.0
info:
  title: Vault API
  version: 0.1.0
  description: |
    Vault HTTP JSON API.

    Conventions:
    - JSON request/response bodies use `application/json` unless otherwise noted.
    - Error responses use the `ErrorResponse` schema.
    - Pagination: list endpoints use `limit` with cursor-based pagination via `cursor`/`nextCursor`. The media list also supports page-based pagination via `page`/`nextPage` when `cursor` is omitted. Max `limit` for media list is 100.
    - IDs are strings; media IDs are validated as UUIDs in path parameters.
    - Timestamps are RFC3339 `date-time` strings.
    - Auth: `requireAuth` prefers `Authorization: Bearer <token>` and falls back to the `access_token` cookie. `/api/auth/me` is cookie-only.
    - Query array params use `style: form` with `explode: true` (e.g., `tags=one&tags=two`).
servers:
  - url: http://127.0.0.1:8000
    description: Local dev
...
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /api/auth/me:
    get:
      summary: Get the current user
      description: Reads the `access_token` cookie.
      tags:
        - Auth
      operationId: auth_me
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthMeResponse"
              examples:
                success:
                  value:
                    user:
                      id: "usr_123"
                      email: "user@example.com"
                      name: "Casey"
                      username: "casey"
                      avatarUrl: "https://cdn.example.com/avatar.png"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/auth/register:
    post:
      summary: Register a new user
      tags:
        - Auth
      operationId: auth_register
      security: []
      x-rateLimit:
        max: 5
        timeWindow: "1 minute"
        keyBy: IP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRegisterRequest"
            examples:
              register:
                value:
                  email: "user@example.com"
                  password: "correct-horse-battery"
      responses:
        "200":
          description: Registration completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthRegisterResponse"
              examples:
                success:
                  value:
                    user:
                      id: "usr_123"
                      email: "user@example.com"
                    access: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.access"
                    refresh: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.refresh"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                badRequest:
                  value:
                    statusCode: 400
                    error: "Bad Request"
                    message: "User already exists"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/auth/login:
    post:
      summary: Log in and set auth cookies
      description: |
        Sets `access_token` (20 h) and `refresh_token` (7 d) httpOnly cookies.
        Rate limited to 30 attempts/min per IP and 10 attempts/min per email.
      tags:
        - Auth
      operationId: auth_login
      security: []
      x-rateLimit:
        - max: 30
          timeWindow: "1 minute"
          keyBy: IP
        - max: 10
          timeWindow: "1 minute"
          keyBy: email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthLoginRequest"
            examples:
              login:
                value:
                  email: "user@example.com"
                  password: "correct-horse-battery"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthLoginResponse"
              examples:
                success:
                  value:
                    user:
                      id: "usr_123"
                      email: "user@example.com"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                badRequest:
                  value:
                    statusCode: 400
                    error: "Bad Request"
                    message: "Invalid email or password format"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                unauthorized:
                  value:
                    statusCode: 401
                    error: "Unauthorized"
                    message: "Invalid credentials"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/auth/refresh:
    post:
      summary: Refresh the access token
      tags:
        - Auth
      operationId: auth_refresh
      security: []
      x-rateLimit:
        max: 60
        timeWindow: "1 minute"
        keyBy: IP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRefreshRequest"
            examples:
              refresh:
                value:
                  refresh: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.refresh"
      responses:
        "200":
          description: Access token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthRefreshResponse"
              examples:
                success:
                  value:
                    access: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.access"
        "401":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                unauthorized:
                  value:
                    statusCode: 401
                    error: "Unauthorized"
                    message: "Invalid or expired refresh token"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ---------------------------------------------------------------------------
  # Profile
  # ---------------------------------------------------------------------------
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token

  # ---------------------------------------------------------------------------
  # Reusable parameters
  # ---------------------------------------------------------------------------
...
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            unauthorized:
              $ref: "#/components/examples/ErrorUnauthorized"
...
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            rateLimited:
              $ref: "#/components/examples/ErrorRateLimited"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            serverError:
              $ref: "#/components/examples/ErrorInternalServer"

  # ---------------------------------------------------------------------------
  # Schemas
  # ---------------------------------------------------------------------------
  schemas:
    ErrorResponse:
      type: object
      description: Error response payload from Fastify and rate limiting.
      properties:
        statusCode:
          type: integer
        error:
          type: string
        message:
          type: string
        limit:
          type: integer
        windowMs:
          type: integer
...
    User:
      type: object
      description: User summary returned by auth endpoints.
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        name:
          type: [string, "null"]
        username:
          type: [string, "null"]
        avatarUrl:
          type: [string, "null"]
      required:
        - id
        - email
    AuthRegisterRequest:
      type: object
      description: Registration credentials.
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
      required:
        - email
        - password
    AuthRegisterResponse:
      type: object
      description: User and tokens returned after registration.
      properties:
        user:
          $ref: "#/components/schemas/User"
        access:
          type: string
        refresh:
          type: string
      required:
        - user
        - access
        - refresh
    AuthLoginRequest:
      type: object
      description: Login credentials.
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
      required:
        - email
        - password
    AuthLoginResponse:
      type: object
      description: User returned after login. Tokens are set as httpOnly cookies.
      properties:
        user:
          $ref: "#/components/schemas/User"
      required:
        - user
    AuthRefreshRequest:
      type: object
      description: Refresh token payload.
      properties:
        refresh:
          type: string
      required:
        - refresh
    AuthRefreshResponse:
      type: object
      description: New access token.
      properties:
        access:
          type: string
      required:
        - access
    AuthMeResponse:
      type: object
      description: Current user response.
      properties:
        user:
          $ref: "#/components/schemas/User"
      required:
        - user

    # --- Profile ---
...
  examples:
    ErrorUnauthorized:
      value:
        statusCode: 401
        error: "Unauthorized"
        message: "Unauthorized"
...
    ErrorRateLimited:
      value:
        error: "rate_limited"
        limit: 30
        windowMs: 60000
    ErrorInternalServer:
      value:
        statusCode: 500
        error: "Internal Server Error"
        message: "Internal Server Error"